BAH-3268 Refactor mv test files
